abstracted and commened functions that process raw crawls into rawcrawl_util
